Allows all hostnames when trusting all certificates
The SNI issue present in the plugin (the wrong certificate is presented when
calling to a Jenkins URL running HTTPS which hosts multiple domains) cannot
be handled by skipping SSL Certificate validation. Therefore a fix allowing
all hostnames when skipping the SSL Certificate is added to circumvent
the hostname check before the certificate is ignored.